Revert "Feat: Make js 내부에서 html templating으로 Header만들기"
This reverts commit 5377440522543de242585fa034390d19d6072c7b.
