--- js/components/Div.js
@@ -1,13 +1,0 @@
-import Element from './Element.js';
-
-export default class extends Element {
-  constructor() {
-    super();
-  }
-  async getElement(classname, children) {
-    return `
-    <div class="${classname}">
-      ${children}
-    </div>`;
-  }
-}
