--- js/components/Element.js
@@ -1,8 +1,0 @@
-export default class {
-  constructor(type) {
-    this.type = type;
-  }
-  async getHtml() {
-    return '';
-  }
-}
