--- js/constant/header-c.js
@@ -1,10 +1,0 @@
-export const NAVIGATION = {
-  CATEGORY_NAMES: [
-    `홈`,
-    `웹툰∙<i class="fas fa-clock"></i>`,
-    `웹소설∙<i class="fas fa-clock"></i>`,
-    `영화`,
-    `방송`,
-    `책`,
-  ],
-};
