+++ style/header/navigation.css
@@ -45,10 +45,10 @@
 }
 
 .c-navigation__button_type_stage > span {
-  background-color: var(--main-yellow);
+  background-color: var(--yellow);
   color: var(--main-black);
   font-weight: 900;
-  padding: 8px 4px 4px 4px;
+  padding: 4px;
 }
 
 .c-navigation__stage-description {
